requirements.txt
pandas
requests
tqdm
beautifulsoup4
PyPDF2
googlesearch-python
openai